Refactoring player logic
Player logic was moved to its own library file. This cleans up the main game file and restricts access to non-public functions
